Refactor: Centralize degToDmsRational function
