+++ src/actions/download-single-photo.js
@@ -1,108 +1,5 @@
 const piexif = require("piexifjs");
 const { Readable } = require("stream");
-
-const { getAuthenticatedClient } = require("../oauth");
-const { downloadPhoto } = require("../photo-manager");
-const {
-  getDriveClient,
-  findOrCreateFolder,
-  createFile,
-  updateFile,
-  findFile,
-  FOLDER_NAME,
-} = require("../drive-manager");
-const { updateState } = require("../download-state");
-
-/**
- * Converts degrees to degrees-minutes-seconds rational format for EXIF data.
- * @param {number} deg - The degree value.
- * @returns {Array<Array<number>>} The DMS rational value.
- */
-function degToDmsRational(deg) {
-    const d = Math.floor(deg);
-    const minFloat = (deg - d) * 60;
-    const m = Math.floor(minFloat);
-    const secFloat = (minFloat - m) * 60;
-    const s = Math.round(secFloat * 100);
-    return [
-      [d, 1],
-      [m, 1],
-      [s, 100],
-    ];
-  }
-
-/**
- * Downloads a single photo to Google Drive.
- * @param {object} req - The Express request object, containing the session.
- * @param {object} photo - The photo object to download.
- */
-async function downloadSinglePhoto(req, photo) {
-    /**
-     * Sends progress updates to the client.
-     * @param {object} progress - The progress object.
-     */
-    const progressCallback = (progress) => {
-      updateState(progress);
-    };
-  
-    try {
-      // Get authenticated clients for Google Drive and Street View
-      const oAuth2Client = await getAuthenticatedClient(req);
-      const drive = await getDriveClient(oAuth2Client);
-
-      // Find or create the folder in Google Drive
-      const folder = await findOrCreateFolder(drive, FOLDER_NAME);
-      const folderId = folder.id;
-  
-      // Send initial progress message
-      progressCallback({
-        message: `Starting download of 1 photo to Google Drive...`,
-        total: 1,
-        current: 0,
-        totalProgress: 0,
-      });
-  
-      const fileName = `${photo.photoId.id}.jpg`;
-  
-      // Send progress update for the current photo
-      progressCallback({
-        message: `Processing photo ${fileName}...`,
-        total: 1,
-        current: 0,
-        photoId: photo.photoId.id,
-        downloadProgress: 0,
-        uploadProgress: 0,
-      });
-  
-      // Download the photo data
-      const { data } = await downloadPhoto(
-        photo.downloadUrl,
-        oAuth2Client,
-        (percentage) => {
-          progressCallback({ downloadProgress: percentage });
-        }
-      );
-  
-      // Add GPS data to the photo's EXIF
-      const jpegData = data.toString("binary");
-      const exifObj = piexif.load(jpegData);
-      const lat = photo.pose.latLngPair.latitude;
-      const lng = photo.pose.latLngPair.longitude;
-      const gpsData = {
-        [piexif.GPSIFD.GPSLatitudeRef]: lat < 0 ? "S" : "N",
-        [piexif.GPSIFD.GPSLatitude]: degToDmsRational(Math.abs(lat)),
-        [piexif.GPSIFD.GPSLongitudeRef]: lng < 0 ? "W" : "E",
-        [piexif.GPSIFD.GPSLongitude]: degToDmsRational(Math.abs(lng)),
-      };
-
-      if (typeof photo.pose.altitude === 'number') {
-        gpsData[piexif.GPSIFD.GPSAltitude] = [Math.round(photo.pose.altitude * 100), 100];
-        gpsData[piexif.GPSIFD.GPSAltitudeRef] = 0;
-      }
-
-      if (typeof photo.pose.heading === 'number') {
-        const piexif = require("piexifjs");
-const { Readable } = require("stream");
 const { exiftool } = require("exiftool-vendored");
 const { getAuthenticatedClient } = require("../oauth");
 const { downloadPhoto } = require("../photo-manager");
@@ -115,24 +12,7 @@ const {
   FOLDER_NAME,
 } = require("../drive-manager");
 const { updateState } = require("../download-state");
-
-/**
- * Converts degrees to degrees-minutes-seconds rational format for EXIF data.
- * @param {number} deg - The degree value.
- * @returns {Array<Array<number>>} The DMS rational value.
- */
-function degToDmsRational(deg) {
-    const d = Math.floor(deg);
-    const minFloat = (deg - d) * 60;
-    const m = Math.floor(minFloat);
-    const secFloat = (minFloat - m) * 60;
-    const s = Math.round(secFloat * 100);
-    return [
-      [d, 1],
-      [m, 1],
-      [s, 100],
-    ];
-  }
+const { degToDmsRational } = require("../utils/photo-utils");
 
 /**
  * Downloads a single photo to Google Drive.
@@ -219,105 +99,6 @@ async function downloadSinglePhoto(req, photo) {
         };
         newJpeg = await exiftool.write(newJpeg, tags);
       }
-  
-      // Create a readable stream from the photo data
-      const stream = Readable.from(newJpeg);
-  
-      // Check if the file already exists in Google Drive
-      const existingFile = await findFile(drive, fileName, folderId);
-  
-      // If the file exists, update it. Otherwise, create a new file.
-      if (existingFile) {
-        await updateFile(
-          drive,
-          existingFile.id,
-          "image/jpeg",
-          stream,
-          newJpeg.length,
-          (percentage) => {
-            progressCallback({ uploadProgress: percentage });
-          }
-        );
-      } else {
-        await createFile(
-          drive,
-          fileName,
-          "image/jpeg",
-          stream,
-          folderId,
-          newJpeg.length,
-          (percentage) => {
-            progressCallback({ uploadProgress: percentage });
-          }
-        );
-      }
-  
-      // Update session data
-      if (req.session.missingPhotos && req.session.downloadedPhotos) {
-        const downloadedPhotoIndex = req.session.missingPhotos.findIndex(
-          (p) => p.photoId.id === photo.photoId.id
-        );
-        if (downloadedPhotoIndex > -1) {
-          const [downloadedPhoto] = req.session.missingPhotos.splice(
-            downloadedPhotoIndex,
-            1
-          );
-          req.session.downloadedPhotos.push(downloadedPhoto);
-        }
-      }
-  
-      // Send progress update
-      progressCallback({
-        fileComplete: true,
-        downloadedCount: req.session.downloadedPhotos.length,
-        notDownloadedCount: req.session.missingPhotos.length,
-        totalProgress: 100,
-      });
-  
-      // Send final progress message
-      progressCallback({
-        message: "Photo downloaded successfully to Google Drive!",
-        complete: true,
-        inProgress: false,
-      });
-    } catch (error) {
-      // Handle errors
-      progressCallback({
-        error: `An error occurred: ${error.message}`,
-        complete: true,
-        inProgress: false,
-      });
-      console.error(error);
-    }
-  }
-
-module.exports = { downloadSinglePhoto };
-
-        gpsData[piexif.GPSIFD.GPSImgDirectionRef] = "T";
-      }
-      exifObj["GPS"] = gpsData;
-      const exifbytes = piexif.dump(exifObj);
-      let newData = piexif.insert(exifbytes, jpegData);
-
-      if (typeof photo.pose.pitch === 'number' || typeof photo.pose.roll === 'number') {
-        const xmpString = `
-          <x:xmpmeta xmlns:x="adobe:ns:meta/" x:xmptk="Adobe XMP Core 5.6-c148 79.164036, 2019/08/13-01:06:57        ">
-            <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#">
-              <rdf:Description rdf:about=""
-                xmlns:GPano="http://ns.google.com/photos/1.0/panorama/">
-                <GPano:PosePitchDegrees>${photo.pose.pitch || 0}</GPano:PosePitchDegrees>
-                <GPano:PoseRollDegrees>${photo.pose.roll || 0}</GPano:PoseRollDegrees>
-              </rdf:Description>
-            </rdf:RDF>
-          </x:xmpmeta>
-        `;
-        const editor = new JpegEditor();
-        editor.load(Buffer.from(newData, "binary"));
-        editor.insertXmp(xmpString);
-        newData = editor.save().toString("binary");
-      }
-
-      const newJpeg = Buffer.from(newData, "binary");
   
       // Create a readable stream from the photo data
       const stream = Readable.from(newJpeg);
